General Added: Fixed: - Fixed casting from math.ceil to int (python 2.7 back support) Updated: ************** Wank! Wank! Wank! Added: Fixed: Updated: ************** Israeli TV! Added: Fixed: Updated: ************** Idan plus! Added: Fixed: Updated:

+++ plugin.video.WankWankWank/resources/lib/Fetchers/hidden/LiveCam/bongcams.py
@@ -222,7 +222,7 @@ class BongCams(PornFetcher):
         page_request = self.get_object_request(category_data) if fetched_request is None else fetched_request
         raw_data = page_request.json()
 
-        return math.ceil(raw_data['online_count'] / self.number_of_videos_per_page)
+        return int(math.ceil(raw_data['online_count'] / self.number_of_videos_per_page))
 
     def get_videos_data(self, page_data):
         """
